==> code/0-globals.R <==
# Global parameters

PATH_DB_RAW <- "/Volumes/ExtremeSSD/marketcheck/db/clean/db"

# PATH_DB <- here::here('data', 'db-sample.parquet') # Sample of original data
PATH_DB <- here::here('data', 'db-full.parquet') # Sample of original data

# Min and max vehicle ages

AGE_YEARS_MIN <- 2
AGE_YEARS_MAX <- 9
